+++ test/support/layout_view.ex
@@ -11,6 +11,7 @@ defmodule Phoenix.LiveViewTest.LayoutView do
     ["LAYOUT", render(assigns.view_module, assigns.view_template, assigns)]
   end
 
+  # TODO use @inner_content when we go phoenix ~> 1.5
   def render("root.html", assigns) do
     ~L"""
     ROOTSTART-<%= @val %>-<%= render(@view_module, @view_template, assigns)%>-ROOTEND
@@ -19,13 +20,13 @@ defmodule Phoenix.LiveViewTest.LayoutView do
 
   def render("live.html", assigns) do
     ~L"""
-    LIVELAYOUTSTART-<%= @val %>-<%= @live_view_module.render(assigns) %>-LIVELAYOUTEND
+    LIVELAYOUTSTART-<%= @val %>-<%= @inner_content %>-LIVELAYOUTEND
     """
   end
 
   def render("live-override.html", assigns) do
     ~L"""
-    LIVEOVERRIDESTART-<%= @val %>-<%= @live_view_module.render(assigns) %>-LIVEOVERRIDEEND
+    LIVEOVERRIDESTART-<%= @val %>-<%= @inner_content %>-LIVEOVERRIDEEND
     """
   end